Use Android API to match numbers more accurately
Match against the call/SMS history using Android API instead of the
old logic based on equality. In this way the same number represented
in different ways (such as national and international formats) will be
recognized as such.

For example, it's common for users who cross national borders often to
store contacts with numbers in international (E.164) format such as
+99-1234567, so the contact can be dialed no matter in which country
the user is at. When the user dials or messages that contact, the number
stored in Android's log is +99-1234567. However, when the user is in
the same country as the contact (country with code +99), and receives a
call from that number, the number is going to be identified as local
number 01234567 instead.

In another more complicated example, countries such as Brazil have
more complicated numbering plans with area codes and the same number
could appear multiple times in the phone log as: +55-99-123456789 or
99-123456789 or 099-123456789 or 123456789 depending on if the user
called or received the call and if they were in the same geographical
area or not.

In the previous logic these numbers were not considered the same,
defeating any "repeated call" or "dialed" rules. This change fixes this
particular bug.

Determining an assumed country code is necessary for the new logic to
work. What the new algorithm does is to gather a number of candidate
country codes from multiple sources:
1. mobile networks the device is currently connected to
2. countries of the active SIM cards
   (requires READ_PHONE_STATE permission, skipped if not granted)